after merge

+++ comms/dht/src/actor.rs
@@ -413,7 +413,7 @@ impl DhtActor {
                 let all_buckets = NodeDistance::get_buckets(num_buckets);
 
                 let mut peer_node_ids = vec![];
-                let mut bucket_number:usize  = bucket.2 as usize;
+                let mut bucket_number: i32 = bucket.2 as i32;
                 let mut bucket_min = bucket.0;
                 let mut bucket_max = bucket.1;
                 while peer_node_ids.len() < propagation_factor && bucket_number >= 0 {
@@ -446,8 +446,8 @@ impl DhtActor {
                         Some(n) => n,
                         None => break,
                     };
-                    bucket_min = all_buckets[bucket_number].0;
-                    bucket_max = all_buckets[bucket_number].1;
+                    bucket_min = all_buckets[bucket_number as usize].0;
+                    bucket_max = all_buckets[bucket_number as usize].1;
                 }
                 Ok(peer_node_ids)
             },
